Revert "Xong preprocess+optimize"
