--- GenericExpertSystem/Attribute.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DataMining
-{
-    /// <summary>
-    /// Attribute class
-    /// </summary>
-    /// Created by VHTHANG{02/05/2021}
-    public class Attribute
-    {
-        #region Declare
-        public bool Enabled { get; set; }
-        public String Name { get; set; }
-        public bool IsClassAttribute { get; set; }
-        #endregion
-    }
-}
