--- GenericExpertSystem/AttributeValue.cs
@@ -1,28 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DataMining
-{
-    /// <summary>
-    /// Attribute values class
-    /// </summary>
-    /// /// Created by VHTHANG{02/05/2021}
-    public class AttributeValue
-    {
-        #region Declare
-        public string Label { get; set; }
-        public int Count { get; set; }
-        public Dictionary<string, int> Statistic { get; set; }
-        #endregion
-
-        #region Constructor
-        public AttributeValue()
-        {
-            Statistic = new Dictionary<string, int>();
-        }
-        #endregion
-    }
-}
